jwt-extended

+++ resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, get_jwt_identity, jwt_optional, fresh_jwt_required
 from flask_restful import Resource, reqparse
 from models.item import *
 
@@ -8,7 +8,7 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="This field cannot be blank!")
     parser.add_argument('store_id', type=int, required=True, help="Every item needs a store id!")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -16,6 +16,7 @@ class Item(Resource):
         else:
             return {'Message': "Item '{}' is not found".format(name)}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"Message": "Item '{}' already exists".format(name)}
@@ -29,7 +30,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"Message": "You have to be an admin."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -50,8 +56,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"Items": [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {"Items": [item['name'] for item in items], 'Message': "More data available if you log in."}, 200
